WARNINGS: Fix all x64/MSVC warnings
WIP HACK [MODULES] fix build

+++ sdk/lib/conutils/outstream.c
@@ -176,7 +176,7 @@ ConWrite(
             }
 
             /* Write everything up to \n */
-            dwNumBytes = ((PCWCH)p - (PCWCH)szStr) * sizeof(WCHAR);
+            dwNumBytes = (DWORD)((PCWCH)p - (PCWCH)szStr) * sizeof(WCHAR);
             WriteFile(Stream->hHandle, szStr, dwNumBytes, &dwNumBytes, NULL);
 
             /*
@@ -260,7 +260,7 @@ ConWrite(
             }
 
             /* Write everything up to \n */
-            dwNumBytes = ((PCCH)p - (PCCH)szStr) * sizeof(CHAR);
+            dwNumBytes = (DWORD)((PCCH)p - (PCCH)szStr) * sizeof(CHAR);
             WriteFile(Stream->hHandle, szStr, dwNumBytes, &dwNumBytes, NULL);
 
             /*
@@ -430,7 +430,7 @@ ConPuts(
 {
     INT Len;
 
-    Len = wcslen(szStr);
+    Len = (INT)wcslen(szStr);
     CON_STREAM_WRITE2(Stream, szStr, Len, Len);
 
     /* Fixup returned length in case of errors */
